[ENH] Consolidate stats unit tests and fix linting errors in tests (#421)
* Consolidates stats unit tests
* Changes to simpler assert for stats errors
* Oof linting
