+++ setup.cfg
@@ -8,7 +8,7 @@ parentdir_prefix =
 
 [flake8]
 max-line-length = 99
-exclude=*build/,tests
+exclude=*build/
 ignore = E126,E402,W504
 per-file-ignores =
     */__init__.py:F401
