+++ tedana/tests/test_decay.py
@@ -14,7 +14,7 @@ from tedana.tests.utils import get_test_data_path
 @pytest.fixture(scope='module')
 def testdata1():
     tes = np.array([14.5, 38.5, 62.5])
-    in_files = [op.join(get_test_data_path(), 'echo{0}.nii.gz'.format(i+1))
+    in_files = [op.join(get_test_data_path(), 'echo{0}.nii.gz'.format(i + 1))
                 for i in range(3)]
     data, _ = io.load_data(in_files, n_echos=len(tes))
     mask, mask_sum = utils.make_adaptive_mask(data, getsum=True)
@@ -64,22 +64,21 @@ def test_fit_decay_ts(testdata1):
 
 def test_smoke_fit_decay():
     """
-    test_smoke_fit_decay tests that the function fit_decay returns reasonable 
-    objects with random inputs in the correct format 
+    test_smoke_fit_decay tests that the function fit_decay returns reasonable
+    objects with random inputs in the correct format
     """
     n_samples = 100
     n_echos = 5
     n_times = 20
     data = np.random.random((n_samples, n_echos, n_times))
     tes = np.random.random((n_echos)).tolist()
-    mask = np.random.randint(2, size=n_samples) # generate binary mask of random 0s and 1s
+    mask = np.random.randint(2, size=n_samples)  # generate binary mask of random 0s and 1s
     masksum = np.random.random((n_samples))
-    fittype='loglin'
-    t2s_limited, s0_limited, t2ss, s0vs, t2s_full, s0_full = me.fit_decay(data, 
-                                                                        tes,
-                                                                        mask,
-                                                                        masksum,
-                                                                        fittype)
+    fittype = 'loglin'
+    t2s_limited, s0_limited, t2ss, s0vs, t2s_full, s0_full = me.fit_decay(
+                                                            data, tes,
+                                                            mask, masksum,
+                                                            fittype)
     assert t2s_limited is not None
     assert s0_limited is not None
     assert t2ss is not None
@@ -90,23 +89,23 @@ def test_smoke_fit_decay():
 
 def test_smoke_fit_decay_curvefit():
     """
-    test_smoke_fit_decay tests that the function fit_decay returns reasonable 
+    test_smoke_fit_decay tests that the function fit_decay returns reasonable
     objects with random inputs in the correct format when using the direct
-    monoexponetial approach 
+    monoexponetial approach
     """
     n_samples = 100
     n_echos = 5
     n_times = 20
     data = np.random.random((n_samples, n_echos, n_times))
     tes = np.random.random((n_echos)).tolist()
-    mask = np.random.randint(2, size=n_samples) # generate binary mask of random 0s and 1s
+    mask = np.random.randint(2, size=n_samples)  # generate binary mask of random 0s and 1s
     masksum = np.random.random((n_samples))
-    fittype='curvefit'
-    t2s_limited, s0_limited, t2ss, s0vs, t2s_full, s0_full = me.fit_decay(data, 
-                                                                        tes,
-                                                                        mask,
-                                                                        masksum,
-                                                                        fittype)
+    fittype = 'curvefit'
+    t2s_limited, s0_limited, t2ss, s0vs, t2s_full, s0_full = me.fit_decay(
+                                                                data, tes,
+                                                                mask,
+                                                                masksum,
+                                                                fittype)
     assert t2s_limited is not None
     assert s0_limited is not None
     assert t2ss is not None
@@ -117,22 +116,22 @@ def test_smoke_fit_decay_curvefit():
 
 def test_smoke_fit_decay_ts():
     """
-    test_smoke_fit_decay_ts tests that the function fit_decay_ts returns reasonable 
-    objects with random inputs in the correct format 
+    test_smoke_fit_decay_ts tests that the function fit_decay_ts returns reasonable
+    objects with random inputs in the correct format
     """
     n_samples = 100
     n_echos = 5
     n_times = 20
     data = np.random.random((n_samples, n_echos, n_times))
     tes = np.random.random((n_echos)).tolist()
-    mask = np.random.randint(2, size=n_samples) # generate binary mask of random 0s and 1s
+    mask = np.random.randint(2, size=n_samples)  # generate binary mask of random 0s and 1s
     masksum = np.random.random((n_samples))
-    fittype='loglin'
-    t2s_limited_ts, s0_limited_ts, t2s_full_ts, s0_full_ts = me.fit_decay_ts(data, 
-                                                                        tes,
-                                                                        mask,
-                                                                        masksum,
-                                                                        fittype)
+    fittype = 'loglin'
+    t2s_limited_ts, s0_limited_ts, t2s_full_ts, s0_full_ts = me.fit_decay_ts(
+                                                                data, tes,
+                                                                mask,
+                                                                masksum,
+                                                                fittype)
     assert t2s_limited_ts is not None
     assert s0_limited_ts is not None
     assert t2s_full_ts is not None
@@ -141,23 +140,23 @@ def test_smoke_fit_decay_ts():
 
 def test_smoke_fit_decay_curvefit_ts():
     """
-    test_smoke_fit_decay_ts tests that the function fit_decay_ts returns reasonable 
+    test_smoke_fit_decay_ts tests that the function fit_decay_ts returns reasonable
     objects with random inputs in the correct format when using the direct
-    monoexponetial approach 
+    monoexponetial approach
     """
     n_samples = 100
     n_echos = 5
     n_times = 20
     data = np.random.random((n_samples, n_echos, n_times))
     tes = np.random.random((n_echos)).tolist()
-    mask = np.random.randint(2, size=n_samples) # generate binary mask of random 0s and 1s
+    mask = np.random.randint(2, size=n_samples)  # generate binary mask of random 0s and 1s
     masksum = np.random.random((n_samples))
-    fittype='curvefit'
-    t2s_limited_ts, s0_limited_ts, t2s_full_ts, s0_full_ts = me.fit_decay_ts(data, 
-                                                                        tes,
-                                                                        mask,
-                                                                        masksum,
-                                                                        fittype)
+    fittype = 'curvefit'
+    t2s_limited_ts, s0_limited_ts, t2s_full_ts, s0_full_ts = me.fit_decay_ts(
+                                                                data, tes,
+                                                                mask,
+                                                                masksum,
+                                                                fittype)
     assert t2s_limited_ts is not None
     assert s0_limited_ts is not None
     assert t2s_full_ts is not None
