+++ tedana/tests/test_gscontrol.py
@@ -18,14 +18,14 @@ def test_break_gscontrol_raw():
     optcom = np.empty((n_samples, n_vols))
     ref_img = ''
 
-    catd = np.empty((n_samples+1, n_echos, n_vols))
+    catd = np.empty((n_samples + 1, n_echos, n_vols))
     with pytest.raises(ValueError) as e_info:
         gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
     assert str(e_info.value) == ('First dimensions of catd ({0}) and optcom ({1}) do not '
                                  'match'.format(catd.shape[0], optcom.shape[0]))
 
-    catd = np.empty((n_samples, n_echos+1, n_vols))
+    catd = np.empty((n_samples, n_echos + 1, n_vols))
     with pytest.raises(ValueError) as e_info:
         gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
@@ -33,7 +33,7 @@ def test_break_gscontrol_raw():
                                  'n_echos ({1})'.format(catd.shape[1], n_echos))
 
     catd = np.empty((n_samples, n_echos, n_vols))
-    optcom = np.empty((n_samples, n_vols+1))
+    optcom = np.empty((n_samples, n_vols + 1))
     with pytest.raises(ValueError) as e_info:
         gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
