+++ tedana/tests/test_model_kundu_metrics.py
@@ -2,7 +2,6 @@
 Tests for tedana.model.fit
 """
 
-import pytest
 import numpy as np
 import pandas as pd
