+++ tedana/tests/test_viz.py
@@ -6,6 +6,7 @@ import numpy as np
 
 from tedana import viz
 
+
 def test_smoke_trim_edge_zeros():
     """
     Ensures that trim_edge_zeros works with random inputs
@@ -14,4 +15,4 @@ def test_smoke_trim_edge_zeros():
 
     assert viz.trim_edge_zeros(arr) is not None
 
-#TODO: Test other functions in viz that generate images?
+# TODO: Test other functions in viz that generate images?
